cloud/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.5
google-cloud-storage==2.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
anthropic==0.7.8
python-dotenv==1.0.0
pydantic==2.5.0
pillow==10.1.0
pillow-heif==0.13.1
